geminis/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
google-cloud-storage>=2.16
pymupdf>=1.24.3
rapidfuzz>=3.9.0
pytesseract>=0.3.10
Pillow>=10.2.0
pydantic>=2.7
psycopg[binary,pool]>=3.2
